refactor: moved suggestisons to common
